+++ packages/frontend/src/data/notifications.ts
@@ -3,7 +3,7 @@ import { common } from "./common";
 const notifications = {
   head: {
     title: "Notifications",
-    description: "Take a look at your notifications.",
+    description: `Take a look at what's happening at the ${common.title}.`,
     author: common.author,
     keywords: ["notifications", "updates"],
     favicon: common.favicon,